+++ common/src/main/java/dev/rinchan/skillrespecpill/mixin/ClientTextTooltipAccessor.java
@@ -1,0 +1,12 @@
+package dev.rinchan.skillrespecpill.mixin;
+
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTextTooltip;
+import net.minecraft.util.FormattedCharSequence;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(ClientTextTooltip.class)
+public interface ClientTextTooltipAccessor {
+    @Accessor("text")
+    FormattedCharSequence skillRespecPill$getText();
+}